requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.6
djangorestframework==3.7.1
httpie==0.9.9
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
